Move deprecated extension functions to a separate file
